Refactor dashboard layout by removing old page styles and integrating components into the main page
