--- src/app/page.module.css
@@ -1,15 +1,0 @@
-.page {
-  
-}
-
-
-
-.main {
- 
-}
-
-
-.footer {
-
-}
-
